refactoring - put the extensions into a library object and each OMCOre now instantiates an extension class
