+++ src/main/scala/diplomaticobjectmodel/model/CustomISAExtensions.scala
@@ -1,6 +1,9 @@
-package diplomaticobjectmodel.model
+// See LICENSE.SiFive for license details.
 
-import freechips.rocketchip.diplomaticobjectmodel.model.{OMEnum, Xsifivecflushdlone}
+package freechips.rocketchip.diplomaticobjectmodel.model
+
+
+import scala.collection.mutable
 
 trait OMCustomExtensionSpecification{
   def name: String
@@ -8,15 +11,35 @@ trait OMCustomExtensionSpecification{
   def _types: Seq[String] = Seq("OMCustomExtensionSpecification", "OMSpecification")
 }
 
-object CustomISAExtensions {
+case class Xsifivecflushdlone(
+  name: String,
+  version: String,
+  override val _types: Seq[String] = Seq("OMXsifivecflushdlone", "OMCustomExtensionSpecification", "OMSpecification")
+) extends OMCustomExtensionSpecification
 
-  trait OMCustomExtensionType extends OMEnum
-  case object XsifivecflushdloneKey extends OMCustomExtensionType
 
-  def cflush: (String) => OMCustomExtensionSpecification = (s: String) => Xsifivecflushdlone("SiFive Extension for Cache Flush", s)
+trait OMCustomExtensionType extends OMEnum
+case object XsifivecflushdloneKey extends OMCustomExtensionType
 
-  val customSpecifications = Map[OMCustomExtensionType, (String) => OMCustomExtensionSpecification](
-    XsifivecflushdloneKey -> cflush
+object CustomExtensionsLibrary {
+
+  private def cflushFunc: (String) => OMCustomExtensionSpecification = (s: String) => Xsifivecflushdlone("SiFive Extension for Cache Flush", s)
+
+  private val lib = Map[OMCustomExtensionType, (String) => OMCustomExtensionSpecification](
+    (XsifivecflushdloneKey -> cflushFunc)
   )
 
+  def get(key: OMCustomExtensionType): (String) => OMCustomExtensionSpecification =
+    lib.get(key).getOrElse(throw new IllegalArgumentException("Error: key: $key not found in custom extension library."))
+}
+
+class CustomISAExtensions {
+  private val customSpecifications = mutable.ArrayBuffer[OMCustomExtensionSpecification]()
+
+  def add(key: OMCustomExtensionType, version: String): Unit = {
+    val f = CustomExtensionsLibrary.get(key)(version)
+    customSpecifications += f
+  }
+
+  def get(): List[OMCustomExtensionSpecification] = customSpecifications.toList
 }
